requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.30.0
pillow>=9.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
pycocoevalcap>=0.1
numpy>=1.20.0
accelerate>=0.20.0
sentencepiece>=0.1.97
protobuf>=3.20.0
tensorboard>=2.10.0
huggingface-hub>=0.16.0 